generalize TicTacToeRoutes into GameRoutes to cover all GameTypes
